feat: add initial raceselector element, test & css
