+++ client/src/components/LiveRacing/RaceSelector.css
@@ -1,0 +1,8 @@
+/* CSS Styling for the RaceSelector Element */
+
+/* Center the spinner element and heading */
+.centertext {
+  display: grid;
+  justify-items: center;
+  margin: 2rem 0;
+}
