+++ client/src/components/LiveRacing/RaceSelector.js
@@ -1,0 +1,15 @@
+// This component will return an Element to select a race from todays races
+
+// Import dependancies
+import Spinner from "../Spinner";
+
+// Render Element
+const RaceSelector = () => {
+  return (
+    <div className="centertext">
+      <h2>RaceSelector Element</h2>
+      <Spinner />
+    </div>
+  );
+};
+export default RaceSelector;
