--- client/src/components/RaceDay/LiveRacing.css
@@ -1,36 +1,0 @@
-/* CSS Styling for the LiveRaceing component */
-
-.meetings-container {
-  display: grid;
-  grid-template-columns: 3fr 2fr 1fr;
-  text-align: left;
-  margin: 1rem;
-  padding: 0.5rem;
-  background-color: var(--background);
-}
-
-/* Center the spinner element and heading */
-.centertext {
-  display: grid;
-  justify-items: center;
-  margin: 2rem 0;
-}
-
-/* Remove list bullet points */
-ol {
-  list-style-type: none;
-}
-
-li {
-  padding: 0.1rem 0;
-}
-
-/* Media query - limit element width for non mobile displays */
-
-@media only screen and (min-width: 51.25rem) {
-  .meetings-container {
-    display: grid;
-    justify-content: end;
-    width: 50%;
-  }
-}
